Seeder Notfound patient json minified

+++ database/migrations/2020_04_22_111323_create_districts_table.php
@@ -16,8 +16,8 @@ class CreateDistrictsTable extends Migration
         Schema::create('districts', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('lat', 10, 7);
-            $table->decimal('lng', 10, 7);
+            $table->double('lat');
+            $table->double('lng');
             $table->integer('infected');
             $table->integer('cured');
             $table->integer('died');
